+++ src/main/java/agency/july/math/Matrix.java
@@ -3,12 +3,12 @@ package agency.july.math;
 public class Matrix {
 
 	// Произведение матрицы на вектор-столбец
-	public static float[] multiplyByVector (float[][] A, float[] V) { // A - матрица (m x n), V - вектор (m),
+	public static double[] multiplyByVector (double[][] A, double[] V) { // A - матрица (m x n), V - вектор (m),
 		int m = V.length;
 		int n = A.length;
 		if (A[0].length != m) throw new Error("The dimensions of the matrix or vector are unacceptable.");
 		
-		float[] resV = new float[n]; // resV - вектор-столбец (n) (результат)
+		double[] resV = new double[n]; // resV - вектор-столбец (n) (результат)
 		for (int i=0; i < n; i++) {
 			resV[i] = 0;
 			for (int j=0; j < m; j++) {
@@ -17,12 +17,12 @@ public class Matrix {
 		}
 		return resV;
 	}
-	public static float[] multiplyByVector (int[][] A, float[] V) { // A - матрица (m x n), V - вектор (m),
+	public static double[] multiplyByVector (int[][] A, double[] V) { // A - матрица (m x n), V - вектор (m),
 		int m = V.length;
 		int n = A.length;
 		if (A[0].length != m) throw new Error("The dimensions of the matrix or vector are unacceptable.");
 		
-		float[] resV = new float[n]; // resV - вектор-столбец (n) (результат)
+		double[] resV = new double[n]; // resV - вектор-столбец (n) (результат)
 		for (int i=0; i < n; i++) {
 			resV[i] = 0;
 			for (int j=0; j < m; j++) {
